pyproject.toml
[project]
name = "alpha-barrier"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autograd>=1.8.0",
    "backtesting>=0.6.4",
    "imbalanced-learn>=0.13.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "keras>=3.10.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas-ta>=0.3.14b0",
    "plotly>=6.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.0",
    "seaborn>=0.13.2",
    "ta-lib>=0.6.4",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.53.1",
]
